make the kanban x-scrollable and and each collumn y-scrollable
